requirements.txt
websockets>=12.0
smartmoneyconcepts>=0.0.14
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas-ta>=0.3.14
